feat: wrap Tern output in container (#5)
Add the Tern output to the container as a final step to provide more copyright and license information to the container users.
Buildx has been dropped for building the final image, as it didn't seem capable of dealing with an image ID as the reference for the start of the final image. The final container is labeled to provide to provide relevant information.